feat: refactor toast/messaging constants
