+++ src/shared/components/toast/toast.styl
@@ -19,17 +19,17 @@
   background: var(--bg);
   color: var(--stroke);
 
-  &--success {
+  &--SUCCESS {
     --bg: var(--green);
     --border: var(--green-border);
     --stroke: var(--green-stroke);
   }
-  &--info {
+  &--INFO {
     --bg: var(--blue);
     --border: var(--blue-border);
     --stroke: var(--blue-stroke);
   }
-  &--error {
+  &--ERROR {
     --bg: var(--red);
     --border: var(--red-border);
     --stroke: var(--red-stroke);
